Add SetupFile, refs 3596
